% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{capm_regression}
\alias{capm_regression}
\title{CAPM regression for various time windows}
\usage{
capm_regression(x, window, freq)
}
\arguments{
\item{x}{: portfolio level data frame}

\item{window}{: months or days}

\item{freq}{: monthly or daily}
}
\value{
beta coefficient
}
\description{
CAPM regression for various time windows
}
