% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute correlation matrix}
\usage{
compute_correlation(
  data,
  use = "complete.obs",
  method = "pearson",
  diagonal = 1,
  quiet = TRUE,
  shave_upper = TRUE
)
}
\arguments{
\item{data}{a matrix or data frame}

\item{use}{see ?correlate}

\item{method}{see ?correlate}

\item{diagonal}{see ?correlate}

\item{quiet}{see ?correlate}

\item{shave_upper}{see ?correlate}
}
\value{
correlation data fram
}
\description{
Compute correlation matrix
}
