% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{compute_persistence}
\alias{compute_persistence}
\title{Compute Persistence}
\usage{
compute_persistence(data, var, tau, probs)
}
\arguments{
\item{data}{}

\item{var}{}

\item{tau}{}

\item{probs}{}
}
\value{

}
\description{
Compute Persistence
}
