% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{construct_univariate_portfolios}
\alias{construct_univariate_portfolios}
\title{Construct univariate portfolios based on beta}
\usage{
construct_univariate_portfolios(
  data,
  n_portfolios = 10,
  nyse_breakpoints = TRUE
)
}
\arguments{
\item{data}{}

\item{n_portfolios}{}

\item{nyse_breakpoints}{}

\item{var}{}
}
\value{

}
\description{
Construct univariate portfolios based on beta
}
\details{
This function takes long data with (stock, month, window)-level
data to construct univariate portfolios.
}
