% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{estimate_portfolio_returns}
\alias{estimate_portfolio_returns}
\title{Estimate portfolio returns}
\usage{
estimate_portfolio_returns(data, ret, n_portfolios = 10, lag = 6)
}
\arguments{
\item{data}{}

\item{ret}{Either \code{ret_ew} or \code{ret_vw}. Note that both variables
are excess returns, including \code{ret_mkt}}

\item{n_portfolios}{}

\item{lag}{}
}
\value{

}
\description{
Estimate portfolio returns
}
