% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{get_portfolio}
\alias{get_portfolio}
\title{Given a list of breakpoints, sort stcoks into portfolio}
\usage{
get_portfolio(x, breakpoints)
}
\arguments{
\item{x}{a value that is sorted into portfolio}

\item{breakpoints}{}
}
\value{

}
\description{
Given a list of breakpoints, sort stcoks into portfolio
}
