% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{quibble}
\alias{quibble}
\title{Returns quantile of each grouped variable}
\usage{
quibble(x, q)
}
\arguments{
\item{x}{variable}

\item{q}{quantile cut points}
}
\value{
data frame
}
\description{
Returns quantile of each grouped variable
}
\details{
Be careful NOT to name the variable when you use this function.
}
\examples{
# OK
as_tibble(mtcars) \%>\% 
  group_by(cyl) \%>\% 
  summarise(quibble(mpg, c(.25, .5, .75)))

# Don't; returns "Data-frame columns"
as_tibble(mtcars) \%>\% 
  group_by(cyl) \%>\% 
  summarise(y = quibble(mpg, c(.25, .5, .75)))

}
