% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{reshape_portfolio_returns}
\alias{reshape_portfolio_returns}
\title{Reshape the result to make it presentable}
\usage{
reshape_portfolio_returns(x, n_portfolios = 10)
}
\arguments{
\item{x}{Output of \code{estimate_portfolio_returns()}}
}
\value{

}
\description{
Reshape the result to make it presentable
}
