% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{roll_rolling_capm_regression}
\alias{roll_rolling_capm_regression}
\title{Rolling capm regressoin using {roll} package}
\usage{
roll_rolling_capm_regression(data, window, freq)
}
\arguments{
\item{window}{}

\item{freq}{}

\item{x}{}
}
\value{

}
\description{
Rolling capm regressoin using {roll} package
}
\details{
We can use {roll} package for monthly data, but not for 
daily data. For daily data, we have to slide within a month, which
makes the use of `slider::slide_period` essential.
}
