% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{rolling_capm_regression}
\alias{rolling_capm_regression}
\title{Rolling capm regressoin for each window and frequency}
\usage{
rolling_capm_regression(x, window, freq)
}
\arguments{
\item{x}{: portfolio level data frame}

\item{window}{: months or days}

\item{freq}{: monthly or daily}
}
\value{
result of the rolling regression
}
\description{
Rolling capm regressoin for each window and frequency
}
