% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Wrapper of \code{weighted.mean()} to deal with NA in weights (\code{w})}
\usage{
weighted_mean(x, w, ..., na.rm = FALSE)
}
\arguments{
\item{x}{}

\item{w}{}

\item{...}{}

\item{na.rm}{}
}
\value{

}
\description{
Wrapper of \code{weighted.mean()} to deal with NA in weights (\code{w})
}
